+++ blocks/product-category-list/product-category-list.js
@@ -1,11 +1,30 @@
 import ffetch from '../../scripts/ffetch.js';
-import { ul } from '../../scripts/dom-builder.js';
-
+import { ul, li, a, h4, span } from '../../scripts/dom-builder.js';
+import { makePublicUrl, imageHelper } from '../../scripts/scripts.js';
 import { toClassName } from '../../scripts/lib-franklin.js';
 import createFilters from './filter.js';
-import createCard from './miniCard.js';
 
 const getSelectionFromUrl = (field) => toClassName(new URLSearchParams(window.location.search).get(field)) || '';
+
+function createCard(product, firstCard = false) {
+  const cardWrapper = a(
+    { href: makePublicUrl(product.path), title: product.title },
+    imageHelper(product.image, product.title, firstCard),
+    h4(
+      {
+        class: '!text-sm !font-normal !text-danahergray-900 !break-words !h-16',
+      },
+      product.title,
+      span({ class: 'text-lg font-semibold text-danaherpurple-500' }, '→'),
+    ),
+  );
+  const card = li({
+    class:
+        'w-52 lg:w-44 h-52 flex flex-col col-span-1 relative mx-auto justify-center transform transition duration-500 border-l-0-5 border-black hover:scale-105 overflow-hidden bg-white max-w-xl pl-6 pr-6 lg:pr-0',
+  }, cardWrapper);
+  card.querySelector('img').className = 'mb-2 h-24 w-full object-cover';
+  return card;
+}
 
 export default async function decorate(block) {
   block.parentElement.parentElement.classList.add('!pb-0');
@@ -35,7 +54,7 @@ export default async function decorate(block) {
   filteredProducts.forEach((product, index) => {
     cardList.append(createCard(product, index === 0));
   });
-  const filterTags = createFilters(products, activeTagFilter);
+  const filterTags = createFilters(products, activeTagFilter, 'brand');
   block.textContent = '';
   block.append(filterTags, cardList);
 }
